Split types.go into multiple files
