+++ Makefile
@@ -189,7 +189,7 @@ presubmit: updateversions regen bin test e2e
 CURRENT_TAG = v0.9.0-gke.1
 #TODO(macxamin) Sync CURRENT_DATASOURCE_SYNCER_TAG with CURRENT_TAG
 CURRENT_DATASOURCE_SYNCER_TAG = v0.9.1-gke.0
-CURRENT_PROM_TAG = v2.41.0-gmp.7-gke.0
+CURRENT_PROM_TAG = v2.41.0-gmp.9-gke.0
 CURRENT_AM_TAG = v0.25.1-gmp.2-gke.0
 LABEL_API_VERSION = 0.9.0
 FILES_TO_UPDATE = $(shell find manifests cmd/operator/deploy -type f -name "*.yaml")
