+++ pkg/operator/apis/monitoring/v1/monitoring_types.go
@@ -1,0 +1,61 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package v1
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// MonitoringConditionType is the type of MonitoringCondition.
+type MonitoringConditionType string
+
+const (
+	// ConfigurationCreateSuccess indicates that the config generated from the
+	// monitoring resource was created successfully.
+	ConfigurationCreateSuccess MonitoringConditionType = "ConfigurationCreateSuccess"
+)
+
+// MonitoringCondition describes a condition of a PodMonitoring.
+type MonitoringCondition struct {
+	Type MonitoringConditionType `json:"type"`
+	// Status of the condition, one of True, False, Unknown.
+	Status corev1.ConditionStatus `json:"status"`
+	// The last time this condition was updated.
+	// +optional
+	LastUpdateTime metav1.Time `json:"lastUpdateTime,omitempty"`
+	// Last time the condition transitioned from one status to another.
+	// +optional
+	LastTransitionTime metav1.Time `json:"lastTransitionTime,omitempty"`
+	// The reason for the condition's last transition.
+	// +optional
+	Reason string `json:"reason,omitempty"`
+	// A human-readable message indicating details about the transition.
+	// +optional
+	Message string `json:"message,omitempty"`
+}
+
+// NewDefaultConditions returns a list of default conditions for at the given
+// time for a `PodMonitoringStatus` if never explicitly set.
+func NewDefaultConditions(now metav1.Time) []MonitoringCondition {
+	return []MonitoringCondition{
+		{
+			Type:               ConfigurationCreateSuccess,
+			Status:             corev1.ConditionUnknown,
+			LastUpdateTime:     now,
+			LastTransitionTime: now,
+		},
+	}
+}
