+++ pkg/operator/apis/monitoring/v1/pod_types.go
@@ -21,6 +21,7 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/export"
 	"github.com/prometheus/common/config"
 	prommodel "github.com/prometheus/common/model"
 	promconfig "github.com/prometheus/prometheus/config"
@@ -28,253 +29,16 @@ import (
 	discoverykube "github.com/prometheus/prometheus/discovery/kubernetes"
 	"github.com/prometheus/prometheus/model/relabel"
 	yaml "gopkg.in/yaml.v2"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-
-	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/export"
 )
 
 var (
 	errInvalidCond = fmt.Errorf("condition needs both 'Type' and 'Status' fields set")
 )
-
-// OperatorConfig defines configuration of the gmp-operator.
-// +genclient
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:storageversion
-type OperatorConfig struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Rules specifies how the operator configures and deployes rule-evaluator.
-	Rules RuleEvaluatorSpec `json:"rules,omitempty"`
-	// Collection specifies how the operator configures collection.
-	Collection CollectionSpec `json:"collection,omitempty"`
-	// ManagedAlertmanager holds information for configuring the managed instance of Alertmanager.
-	// +kubebuilder:default={configSecret: {name: alertmanager, key: alertmanager.yaml}}
-	ManagedAlertmanager *ManagedAlertmanagerSpec `json:"managedAlertmanager,omitempty"`
-	// Features holds configuration for optional managed-collection features.
-	Features OperatorFeatures `json:"features,omitempty"`
-}
-
-// OperatorConfigList is a list of OperatorConfigs.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type OperatorConfigList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []OperatorConfig `json:"items"`
-}
-
-// RuleEvaluatorSpec defines configuration for deploying rule-evaluator.
-type RuleEvaluatorSpec struct {
-	// ExternalLabels specifies external labels that are attached to any rule
-	// results and alerts produced by rules. The precedence behavior matches that
-	// of Prometheus.
-	ExternalLabels map[string]string `json:"externalLabels,omitempty"`
-	// QueryProjectID is the GCP project ID to evaluate rules against.
-	// If left blank, the rule-evaluator will try attempt to infer the Project ID
-	// from the environment.
-	QueryProjectID string `json:"queryProjectID,omitempty"`
-	// The base URL used for the generator URL in the alert notification payload.
-	// Should point to an instance of a query frontend that gives access to queryProjectID.
-	GeneratorURL string `json:"generatorUrl,omitempty"`
-	// Alerting contains how the rule-evaluator configures alerting.
-	Alerting AlertingSpec `json:"alerting,omitempty"`
-	// A reference to GCP service account credentials with which the rule
-	// evaluator container is run. It needs to have metric read permissions
-	// against queryProjectId and metric write permissions against all projects
-	// to which rule results are written.
-	// Within GKE, this can typically be left empty if the compute default
-	// service account has the required permissions.
-	Credentials *corev1.SecretKeySelector `json:"credentials,omitempty"`
-}
-
-// CollectionSpec specifies how the operator configures collection of metric data.
-type CollectionSpec struct {
-	// ExternalLabels specifies external labels that are attached to all scraped
-	// data before being written to Cloud Monitoring. The precedence behavior matches that
-	// of Prometheus.
-	ExternalLabels map[string]string `json:"externalLabels,omitempty"`
-	// Filter limits which metric data is sent to Cloud Monitoring.
-	Filter ExportFilters `json:"filter,omitempty"`
-	// A reference to GCP service account credentials with which Prometheus collectors
-	// are run. It needs to have metric write permissions for all project IDs to which
-	// data is written.
-	// Within GKE, this can typically be left empty if the compute default
-	// service account has the required permissions.
-	Credentials *corev1.SecretKeySelector `json:"credentials,omitempty"`
-	// Configuration to scrape the metric endpoints of the Kubelets.
-	KubeletScraping *KubeletScraping `json:"kubeletScraping,omitempty"`
-	// Compression enables compression of metrics collection data
-	Compression CompressionType `json:"compression,omitempty"`
-}
-
-// OperatorFeatures holds configuration for optional managed-collection features.
-type OperatorFeatures struct {
-	// Configuration of target status reporting.
-	TargetStatus TargetStatusSpec `json:"targetStatus,omitempty"`
-	// Settings for the collector configuration propagation.
-	Config ConfigSpec `json:"config,omitempty"`
-}
-
-// ConfigSpec holds configurations for the Prometheus configuration.
-type ConfigSpec struct {
-	// Compression enables compression of the config data propagated by the operator to collectors.
-	// It is recommended to use the gzip option when using a large number of ClusterPodMonitoring
-	// and/or PodMonitoring.
-	Compression CompressionType `json:"compression,omitempty"`
-}
-
-// TargetStatusSpec holds configuration for target status reporting.
-type TargetStatusSpec struct {
-	// Enable target status reporting.
-	Enabled bool `json:"enabled,omitempty"`
-}
-
-// +kubebuilder:validation:Enum=none;gzip
-type CompressionType string
-
-const CompressionNone CompressionType = "none"
-const CompressionGzip CompressionType = "gzip"
-
-// KubeletScraping allows enabling scraping of the Kubelets' metric endpoints.
-type KubeletScraping struct {
-	// The interval at which the metric endpoints are scraped.
-	Interval string `json:"interval"`
-}
-
-// ExportFilters provides mechanisms to filter the scraped data that's sent to GMP.
-type ExportFilters struct {
-	// A list Prometheus time series matchers. Every time series must match at least one
-	// of the matchers to be exported. This field can be used equivalently to the match[]
-	// parameter of the Prometheus federation endpoint to selectively export data.
-	// Example: `["{job!='foobar'}", "{__name__!~'container_foo.*|container_bar.*'}"]`
-	MatchOneOf []string `json:"matchOneOf,omitempty"`
-}
-
-// AlertingSpec defines alerting configuration.
-type AlertingSpec struct {
-	// Alertmanagers contains endpoint configuration for designated Alertmanagers.
-	Alertmanagers []AlertmanagerEndpoints `json:"alertmanagers,omitempty"`
-}
-
-// ManagedAlertmanagerSpec holds configuration information for the managed
-// Alertmanager instance.
-type ManagedAlertmanagerSpec struct {
-	// ConfigSecret refers to the name of a single-key Secret in the public namespace that
-	// holds the managed Alertmanager config file.
-	ConfigSecret *corev1.SecretKeySelector `json:"configSecret,omitempty"`
-	// ExternalURL is the URL under which Alertmanager is externally reachable
-	// (for example, if Alertmanager is served via a reverse proxy).
-	// Used for generating relative and absolute links back to Alertmanager
-	// itself. If the URL has a path portion, it will be used to prefix all HTTP
-	// endpoints served by Alertmanager.
-	// If omitted, relevant URL components will be derived automatically.
-	ExternalURL string `json:"externalURL,omitempty"`
-}
-
-// AlertmanagerEndpoints defines a selection of a single Endpoints object
-// containing alertmanager IPs to fire alerts against.
-type AlertmanagerEndpoints struct {
-	// Namespace of Endpoints object.
-	Namespace string `json:"namespace"`
-	// Name of Endpoints object in Namespace.
-	Name string `json:"name"`
-	// Port the Alertmanager API is exposed on.
-	Port intstr.IntOrString `json:"port"`
-	// Scheme to use when firing alerts.
-	Scheme string `json:"scheme,omitempty"`
-	// Prefix for the HTTP path alerts are pushed to.
-	PathPrefix string `json:"pathPrefix,omitempty"`
-	// TLS Config to use for alertmanager connection.
-	TLS *TLSConfig `json:"tls,omitempty"`
-	// Authorization section for this alertmanager endpoint
-	Authorization *Authorization `json:"authorization,omitempty"`
-	// Version of the Alertmanager API that rule-evaluator uses to send alerts. It
-	// can be "v1" or "v2".
-	APIVersion string `json:"apiVersion,omitempty"`
-	// Timeout is a per-target Alertmanager timeout when pushing alerts.
-	Timeout string `json:"timeout,omitempty"`
-}
-
-// Authorization specifies a subset of the Authorization struct, that is
-// safe for use in Endpoints (no CredentialsFile field).
-type Authorization struct {
-	// Set the authentication type. Defaults to Bearer, Basic will cause an
-	// error
-	Type string `json:"type,omitempty"`
-	// The secret's key that contains the credentials of the request
-	Credentials *corev1.SecretKeySelector `json:"credentials,omitempty"`
-}
-
-// Auth sets the `Authorization` header on every scrape request.
-//
-// Currently the credentials are not configurable and always empty.
-type Auth struct {
-	// The authentication type. Defaults to Bearer, Basic will cause an error.
-	Type string `json:"type,omitempty"`
-	// TODO: Add credentials: https://github.com/GoogleCloudPlatform/prometheus-engine/issues/450
-}
-
-// BasicAuth sets the `Authorization` header on every scrape request with the
-// configured username.
-//
-// Currently the password is not configurable and always empty.
-type BasicAuth struct {
-	// The username for authentication.
-	Username string `json:"username,omitempty"`
-	// TODO: Add password: https://github.com/GoogleCloudPlatform/prometheus-engine/issues/450
-}
-
-// TLS specifies TLS configuration parameters from Kubernetes resources.
-type TLS struct {
-	// Used to verify the hostname for the targets.
-	ServerName string `json:"serverName,omitempty"`
-	// Disable target certificate validation.
-	InsecureSkipVerify bool `json:"insecureSkipVerify,omitempty"`
-	// Minimum TLS version. Accepted values: TLS10 (TLS 1.0), TLS11 (TLS 1.1), TLS12 (TLS 1.2), TLS13 (TLS 1.3).
-	// If unset, Prometheus will use Go default minimum version, which is TLS 1.2.
-	// See MinVersion in https://pkg.go.dev/crypto/tls#Config.
-	MinVersion string `json:"minVersion,omitempty"`
-	// Maximum TLS version. Accepted values: TLS10 (TLS 1.0), TLS11 (TLS 1.1), TLS12 (TLS 1.2), TLS13 (TLS 1.3).
-	// If unset, Prometheus will use Go default minimum version, which is TLS 1.2.
-	// See MinVersion in https://pkg.go.dev/crypto/tls#Config.
-	MaxVersion string `json:"maxVersion,omitempty"`
-}
-
-// TLSConfig specifies TLS configuration parameters from Kubernetes resources.
-type TLSConfig struct {
-	// Struct containing the CA cert to use for the targets.
-	CA *SecretOrConfigMap `json:"ca,omitempty"`
-	// Struct containing the client cert file for the targets.
-	Cert *SecretOrConfigMap `json:"cert,omitempty"`
-	// Secret containing the client key file for the targets.
-	KeySecret *corev1.SecretKeySelector `json:"keySecret,omitempty"`
-	// Used to verify the hostname for the targets.
-	ServerName string `json:"serverName,omitempty"`
-	// Disable target certificate validation.
-	InsecureSkipVerify bool `json:"insecureSkipVerify,omitempty"`
-	// Minimum TLS version. Accepted values: TLS10 (TLS 1.0), TLS11 (TLS 1.1), TLS12 (TLS 1.2), TLS13 (TLS 1.3).
-	// If unset, Prometheus will use Go default minimum version, which is TLS 1.2.
-	// See MinVersion in https://pkg.go.dev/crypto/tls#Config.
-	MinVersion string `json:"minVersion,omitempty"`
-	// Maximum TLS version. Accepted values: TLS10 (TLS 1.0), TLS11 (TLS 1.1), TLS12 (TLS 1.2), TLS13 (TLS 1.3).
-	// If unset, Prometheus will use Go default minimum version, which is TLS 1.2.
-	// See MinVersion in https://pkg.go.dev/crypto/tls#Config.
-	MaxVersion string `json:"maxVersion,omitempty"`
-}
-
-// SecretOrConfigMap allows to specify data as a Secret or ConfigMap. Fields are mutually exclusive.
-// Taking inspiration from prometheus-operator: https://github.com/prometheus-operator/prometheus-operator/blob/2c81b0cf6a5673e08057499a08ddce396b19dda4/Documentation/api.md#secretorconfigmap
-type SecretOrConfigMap struct {
-	// Secret containing data to use for the targets.
-	Secret *corev1.SecretKeySelector `json:"secret,omitempty"`
-	// ConfigMap containing data to use for the targets.
-	ConfigMap *corev1.ConfigMapKeySelector `json:"configMap,omitempty"`
-}
 
 // PodMonitoringCRD represents a Kubernetes CRD that monitors Pod endpoints.
 type PodMonitoringCRD interface {
@@ -1100,24 +864,6 @@ type ScrapeEndpoint struct {
 	HTTPClientConfig `json:",inline"`
 }
 
-type ProxyConfig struct {
-	// HTTP proxy server to use to connect to the targets. Encoded passwords are not supported.
-	ProxyURL string `json:"proxyUrl,omitempty"`
-	// TODO(TheSpiritXIII): https://prometheus.io/docs/prometheus/latest/configuration/configuration/#oauth2
-}
-
-// HTTPClientConfig stores HTTP-client configurations.
-type HTTPClientConfig struct {
-	// The HTTP authorization credentials for the targets.
-	Authorization *Auth `json:"authorization,omitempty"`
-	// The HTTP basic authentication credentials for the targets.
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// Configures the scrape request's TLS settings.
-	TLS *TLS `json:"tls,omitempty"`
-	// Proxy configuration.
-	ProxyConfig `json:",inline"`
-}
-
 // TargetLabels configures labels for the discovered Prometheus targets.
 type TargetLabels struct {
 	// Pod metadata labels that are set on all scraped targets.
@@ -1212,209 +958,6 @@ type PodMonitoringStatus struct {
 	Conditions []MonitoringCondition `json:"conditions,omitempty"`
 	// Represents the latest available observations of target state for each ScrapeEndpoint.
 	EndpointStatuses []ScrapeEndpointStatus `json:"endpointStatuses,omitempty"`
-}
-
-// MonitoringConditionType is the type of MonitoringCondition.
-type MonitoringConditionType string
-
-const (
-	// ConfigurationCreateSuccess indicates that the config generated from the
-	// monitoring resource was created successfully.
-	ConfigurationCreateSuccess MonitoringConditionType = "ConfigurationCreateSuccess"
-)
-
-// MonitoringCondition describes a condition of a PodMonitoring.
-type MonitoringCondition struct {
-	Type MonitoringConditionType `json:"type"`
-	// Status of the condition, one of True, False, Unknown.
-	Status corev1.ConditionStatus `json:"status"`
-	// The last time this condition was updated.
-	// +optional
-	LastUpdateTime metav1.Time `json:"lastUpdateTime,omitempty"`
-	// Last time the condition transitioned from one status to another.
-	// +optional
-	LastTransitionTime metav1.Time `json:"lastTransitionTime,omitempty"`
-	// The reason for the condition's last transition.
-	// +optional
-	Reason string `json:"reason,omitempty"`
-	// A human-readable message indicating details about the transition.
-	// +optional
-	Message string `json:"message,omitempty"`
-}
-
-// NewDefaultConditions returns a list of default conditions for at the given
-// time for a `PodMonitoringStatus` if never explicitly set.
-func NewDefaultConditions(now metav1.Time) []MonitoringCondition {
-	return []MonitoringCondition{
-		{
-			Type:               ConfigurationCreateSuccess,
-			Status:             corev1.ConditionUnknown,
-			LastUpdateTime:     now,
-			LastTransitionTime: now,
-		},
-	}
-}
-
-// Rules defines Prometheus alerting and recording rules that are scoped
-// to the namespace of the resource. Only metric data from this namespace is processed
-// and all rule results have their project_id, cluster, and namespace label preserved
-// for query processing.
-// If the location label is not preserved by the rule, it defaults to the cluster's location.
-//
-// +genclient
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:subresource:status
-// +kubebuilder:storageversion
-type Rules struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of rules to record and alert on.
-	Spec RulesSpec `json:"spec"`
-	// Most recently observed status of the resource.
-	// +optional
-	Status RulesStatus `json:"status"`
-}
-
-// RulesList is a list of Rules.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type RulesList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []Rules `json:"items"`
-}
-
-// ClusterRules defines Prometheus alerting and recording rules that are scoped
-// to the current cluster. Only metric data from the current cluster is processed
-// and all rule results have their project_id and cluster label preserved
-// for query processing.
-// If the location label is not preserved by the rule, it defaults to the cluster's location.
-//
-// +genclient
-// +genclient:nonNamespaced
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:resource:scope=Cluster
-// +kubebuilder:subresource:status
-// +kubebuilder:storageversion
-type ClusterRules struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of rules to record and alert on.
-	Spec RulesSpec `json:"spec"`
-	// Most recently observed status of the resource.
-	// +optional
-	Status RulesStatus `json:"status"`
-}
-
-// ClusterRulesList is a list of ClusterRules.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type ClusterRulesList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []ClusterRules `json:"items"`
-}
-
-// GlobalRules defines Prometheus alerting and recording rules that are scoped
-// to all data in the queried project.
-// If the project_id or location labels are not preserved by the rule, they default to
-// the values of the cluster.
-//
-// +genclient
-// +genclient:nonNamespaced
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:resource:scope=Cluster
-// +kubebuilder:subresource:status
-// +kubebuilder:storageversion
-type GlobalRules struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of rules to record and alert on.
-	Spec RulesSpec `json:"spec"`
-	// Most recently observed status of the resource.
-	// +optional
-	Status RulesStatus `json:"status"`
-}
-
-// GlobalRulesList is a list of GlobalRules.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type GlobalRulesList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []GlobalRules `json:"items"`
-}
-
-// RulesSpec contains specification parameters for a Rules resource.
-type RulesSpec struct {
-	// A list of Prometheus rule groups.
-	Groups []RuleGroup `json:"groups"`
-}
-
-// RuleGroup declares rules in the Prometheus format:
-// https://prometheus.io/docs/prometheus/latest/configuration/recording_rules/
-type RuleGroup struct {
-	// The name of the rule group.
-	Name string `json:"name"`
-	// The interval at which to evaluate the rules. Must be a valid Prometheus duration.
-	Interval string `json:"interval"`
-	// A list of rules that are executed sequentially as part of this group.
-	Rules []Rule `json:"rules"`
-}
-
-// Rule is a single rule in the Prometheus format:
-// https://prometheus.io/docs/prometheus/latest/configuration/recording_rules/
-type Rule struct {
-	// Record the result of the expression to this metric name.
-	// Only one of `record` and `alert` must be set.
-	Record string `json:"record,omitempty"`
-	// Name of the alert to evaluate the expression as.
-	// Only one of `record` and `alert` must be set.
-	Alert string `json:"alert,omitempty"`
-	// The PromQL expression to evaluate.
-	Expr string `json:"expr"`
-	// The duration to wait before a firing alert produced by this rule is sent to Alertmanager.
-	// Only valid if `alert` is set.
-	For string `json:"for,omitempty"`
-	// A set of labels to attach to the result of the query expression.
-	Labels map[string]string `json:"labels,omitempty"`
-	// A set of annotations to attach to alerts produced by the query expression.
-	// Only valid if `alert` is set.
-	Annotations map[string]string `json:"annotations,omitempty"`
-}
-
-// RulesStatus contains status information for a Rules resource.
-type RulesStatus struct {
-	// TODO: add status information.
-}
-
-// NodeMonitoringSpec contains specification parameters for NodeMonitoring.
-type NodeMonitoringSpec struct {
-	// Label selector that specifies which nodes are selected for this monitoring
-	// configuration. If left empty all nodes are selected.
-	Selector metav1.LabelSelector `json:"selector,omitempty"`
-	// The endpoints to scrape on the selected nodes.
-	Endpoints []ScrapeEndpoint `json:"endpoints"`
-	// Limits to apply at scrape time.
-	Limits *ScrapeLimits `json:"limits,omitempty"`
-}
-
-// NodeMonitoringList is a list of NodeMonitorings.
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type NodeMonitoringList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []NodeMonitoring `json:"items"`
-}
-
-// NodeMonitoring defines monitoring for a set of nodes.
-// +genclient
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:subresource:status
-// +kubebuilder:storageversion
-type NodeMonitoring struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired node selection for target discovery by
-	// Prometheus.
-	Spec NodeMonitoringSpec `json:"spec"`
 }
 
 var invalidLabelCharRE = regexp.MustCompile(`[^a-zA-Z0-9_]`)
